Add batch notification function
This commit adds an Azure function which calls a simple command class.
The command class creates a json payload from the data it is passed then
calls the Notify API batch message endpoint.
Currently the API response is passed back to the Azure function, we
should wrap this in our own response.