Add azure function which interacts with NHS notify
This azure function is triggered via HTTP and accepts json post data of
a specific format. The data is turned into mulitple requests to the
Notify send message endpoint. An individual routing plan may be
specified per recipient or a top level routing plan can be applied for
all recipients.

Currently responds with a list of send message response text from
Notify, we will decorate or wrap these in a later feature.