Refactor access token, env vars, module names (#17)
* Rename base url env var

We call a specific url for tokens and another for notify, make the env var names reflect the value.

* Refactor how we fetch an access token

We've stubbed most of this out to date, there are secrets we will need from the env to successfully request an access token for NHS Notify. This commit adds some test coverage to the underlying jwt logic and uses more descriptive env var names.

* Add required env var names

* Add local env files to .gitignore

* Make docker compose use env vars

* Temporarily disable Notify access token lookup

This won't work until we've completed Notify integration.

* Use the function name for top level module

This makes the various helpers and test files more distinct from one another and fixes an issue with running the entire test suite.