Add an end to end test
Adds a simple end to end test which runs docker in a subprocess, pauses then performs a file upload in another subprocess.
When containers are stopped the output from the docker subprocess is used to in test assertions.

Adds and example.csv upload file which contains dummy but valid NHS
numbers and no PII.