Add basic end to end test (#20)
* Add a script to upload files to azurite

When running functions and azurite locally this script gives us a
convenient way to upload pilot data.

* Add an end to end test

Adds a simple end to end test which runs docker in a subprocess, pauses then performs a file upload in another subprocess.
When containers are stopped the output from the docker subprocess is used to in test assertions.

Adds and example.csv upload file which contains dummy but valid NHS
numbers and no PII.