Add tests init.py files
These __init__.py files allow the tests to locate the correct functions and associated modules.
It's also necessary to remove the non-unique 'function_app' module from each test package.
This clash occurs because of the limitations of Azure function_app.py naming, we can't nest them in a module heirarchy.