Run end to end tests in their own container (#51)
Adds a bespoke end to end test container which depends on the successful completion of azurite setup. This means the azurite blob containers are present and writable.
The e2e test container shares logs with the function apps which allows it to poll for expected log output indicating the functions have been triggered and the execution was successful.