Save message status (#48)
* Rename test env file

We will use this for the integration and end to end tests.

* Add psycopg2-binary postgres adapter for Python

* Rework database schema

We still need this to dynamically create development, test and production databases.

* Make message-status function use pipenv for dependencies

* Add a shared data storing module

This adds integration tests to work with the database container. Currently this reuses the same postgres db.
The data storing module can be shared between any functions needing to persist message status data.

* Include integration tests in test suite
We might want to break these up a bit so we can run unit, integration and end to end tests independently on CI but for now run integration tests by default.

* Increase polling time for end to end test logs

* Ensure the integration test script returns the exit code from the test run