Initialise database (#105)
* Rename db init script to schema.sql

* Remove db init script auto execution from compose

<sigh> we have had to push this into the application runtime. Yes that's right, you are actually reading this statement.
This is something we need to and will fix with a proper migration build step. Time constraints mean we have to do this.

* Check and initialise schema on connection

I don't like it either. Time constraints have meant we have to push schema creation into the datastore module. It wasn't my idea.
We should iterate this so that we have migrations as part of the deployment pipeline.