Make `messageReference` unique (#118)
* Remove message_reference method

The same functionality will be implemented in recipient_id - a repeatable hash of appt. time, date, NiN and DoB.

* Add a method to generate a uuid4 string

Slightly easier to patch in testing.

* Generate message reference before notifying recipient

We want a unique identifier for each recipient, we need to pass this to the messsage body we send to NHS Notify and to the data we persist for each recipient.

* Update message body in pact tests

* Fix integration test expectations around message reference