Refactor schema for batch message endpoint
We will be sending messages using NHS Notify's message batch endpoint, this means a revision of how we store incoming data and the success of the requests to Notify.
We will be archiving existing data from the pilot, likely by renaming the existing tables so it is fine to start the migration revisions with the intended schema.