Add a way to record message batches
We will be requesting the message-batch endpoint on NHS Notify and this will mean how we store the message batch data will be different to the pilot.
We will record one message batch record to many message records and merge the response id into the original personalisation data in the messages records.