Add CSV file processor to function app
We don't want to lose the functionality from the BSO Pilot (i.e. CSV
uploads calling Notify) but we now do this via the NHS Notify message batch endpoint.
We also do not need the overhead of multiple function apps when the file
upload function can be added to the existing function app and call the
underlying dispatcher service.