Feature Branch: Flask running in function app (#137)
* Remove code not currently in use on feature branch

* Add Flask dependencies

* Add basic Flask app

* Add function app which delegates requests to Flask

* Test Flask app in CI

* Revise Dockerfile/compose.yml for one function app with db

* Add pytest-sugar

We all love a bar chart

* Add request verifier

Uses HMAC signature validation to verify expected headers with a request body.
The actual implementation is abstracted from requests themselves so we deal with strings, dictionaries and booleans as inputs and outputs.

* Access headers with lowercase keys

Flask titlecases the keys in Headers, make sure we lowercase before attempting access.

* Add POST status create endpoint

* Add jsonschema validation library

* Validate request body data against NHS Notify schema

Use the OAS file provided by NHS Notify to validate the incoming POST body.

* Move status endpoints to route handlers

Make healthcheck a top level endpoint, it should not be nested in the
sttaus/ route.

* Update name of healthcheck route function to match path

Co-authored-by: Nimmo <dnimmo@gmail.com>

* Ensure local test runs use test db

* Add status recorder service

When the status callback endpoint receives requests the post body will be saved to the database to record the channel and message status depending on the payload type.
This commit adds the datastore and service module which handles this persistence.

* Use psycopg2 sql.Identifier to safely interpolate the table name

There's good support for composition of SQL in psycopg2 so use it.

* Add FLASK_DEBUG env vars

* Extract hmac signature generation

We do this in test setup quite a few times so make it a utility.

* Read the post body as json and ensure key sorting is consistent

* Truncate all tables after each test

* Return useful error descriptions in the status/create response

* Add postgres to CI test jobs

* Refactor persistence tests

We now have local dev and test dbs so improve tests by asserting persistence has worked as expected.

* alembic init

* added alembic for migrations

* Add Flask dependencies

* Add basic Flask app

* Add function app which delegates requests to Flask

* alembic init

* added alembic for migrations

* PR comments

* file formatting fixes for auto-generated alembic init

* removed schema.hcl

* re-add schema.hcl

* DTOSS-6505 update Pipfile

* replace print with logger

* Simplify db setup

DB creation is handled nicely in the Alembic script, so remove from CI workflow steps and local scripts.

* Add sqlalchemy_utils dependency

* Add alembic_postgresql_enum dependency

* Refactor schema for batch message endpoint

We will be sending messages using NHS Notify's message batch endpoint, this means a revision of how we store incoming data and the success of the requests to Notify.
We will be archiving existing data from the pilot, likely by renaming the existing tables so it is fine to start the migration revisions with the intended schema.

* Remove schema initialisation from database module

This was a temporary measure and a good thing that we are removing it to replace with database migrations.

* Run alembic upgrade head on test CI dbs

* Add database migration to dev environment startup script

* Add a way to record message batches

We will be requesting the message-batch endpoint on NHS Notify and this will mean how we store the message batch data will be different to the pilot.
We will record one message batch record to many message records and merge the response id into the original personalisation data in the messages records.

* Add JWT access token utility module

* Generate a sensible recipient_id per message

* Handle a failed response from Notify when recording a message batch

In this case we want a record of what failed and what the response from Notify was.

* Make a message batch request to NHS Notify and record the outcome

* Add message batch API endpoint

Adds /api/message/batch which follows the same HMAC verification and request body schema validation as with status creation.

* Remove datastore functions to create message batch records

This is now done via SQLAlchemy in message_batch_recorder.

* Replace SQL persistence with SQLAlchemy ORM functions

* Remove unused import

* Fix models import issue in function app

We can't include code from outside the function app root directory without some sort of file moving setup.
Models are now a core part of how we persist and retrieve records from the database, so this commit moves the models definition into the function app within the app module namespace.

* Pass environment to function app when running locally

* Remove schema COPY command from Dockerfile

* Update developer documentation

* Use unique HMAC keys and secrets

The callback requests we receive from NHS Notify should verify the HMAC signature using the Notify application ID and API key.
The requests we receive to send message batches should verify the HMAC signature using client application ID and API key.
We will need to provide a way to manage multiple client configurations, but this commit at least separates the two types of verification.

* Lookup schema in schema validator

We determine the type of request and consequently the schema to validate
with by inspecting the request body, this commit moves the
request/schema type determination to the schema validator for better
encapsulation.
There were a few tests missing for batch message schema validation,
added these too.

* Fix bug with how JSONB fields are saved

We were dumping to JSON string then saving in a JSONB field then loading back into a dict. We don't need to do this, as SQLAlchemy/Psycopg handles this for us.

* Add a service which allows us to query channel statuses

* Add endpoint to query channel status data

Adds a GET endpoint to return statuses found by status_reporter module.
Query params are not validated yet.

* Add ability to search statuses by NHS Number

* Add CSV file processor to function app

We don't want to lose the functionality from the BSO Pilot (i.e. CSV
uploads calling Notify) but we now do this via the NHS Notify message batch endpoint.
We also do not need the overhead of multiple function apps when the file
upload function can be added to the existing function app and call the
underlying dispatcher service.

* Fix import deprecation warning for declarative base

* Save supplier status to channel statuses status column

We usually query this value much more than channel status. We also want to ensure the values conform to the define db enum.
This commit also changes the supplierStatus value on the default test channel status post body fixture. So some tests are also adjusted to reflect this.

* Add a foreign key to status tables to messages.message_id

It doesn't makes sense to persist status data without a corresponding message.

* Fix batch recording bug

Unit tests did not accurately reflect the request body when saving message batches to the db.

* Fix a couple of rebase issues

This branch has now been rebased with main and required a couple of small fixes to issues occurring from rebase resolution.

---------

Co-authored-by: Nimmo <dnimmo@gmail.com>