Revert tagging images from PR merges with integration
The default environment tag is development, which means PR merges to main will tag the resulting image with this value. This image will then be available on the corresponding Azure environment, and we can then manage the re-tagging of these images using the devops pipeline. This gives us a way to (almost) automatically deploy to our dev env, re-tag to integration and/or nft and then out to preprod and production.