Add end to end test for status callbacks and GET statuses endpoint (#178)
* Make scope for truncating db dynamic

This can be session, module or function, defaults to function.

* Make callbacks from NHS Notify API Stub

We want to emulate a real set of message and channel status callbacks from NHS Notify API so that we can test the status endpoints in our app.

* Add pytest_steps

End to end tests get a bit monolithic, pytest steps breaks down the output into logical steps and halts on a step failure.

* Add tests for status endpoints

We make fake callbacks from the NHS Notify API Stub, this tests they save data correctly and that the data is available via the GET statuses endpoint.