Include the client bearer token if present in headers (#207)
NHS Notify need to identify requests from clients to ensure the correct
routing plans are in use.
This commit allows the incoming Authorization header value to be
preserved and used in the request made to the NHS Notify API, otherwise
the default credentials are used to generate the JWT / Bearer token.
This fallback preserves existing BSO file upload functionality.