Prevent missing bearer token via batch message endpoint
We expect clients to supply a valid JWT bearer token in request headers, if it isn't present ensure we populate an invalid token.
This prevents a fallback to using internal credentials to generate a token as we only do this on file upload which is secured on Azure blob storage.