Remove hmac verification on message batch endpoint (#213)
* Prevent missing bearer token via batch message endpoint

We expect clients to supply a valid JWT bearer token in request headers, if it isn't present ensure we populate an invalid token.
This prevents a fallback to using internal credentials to generate a token as we only do this on file upload which is secured on Azure blob storage.

* Check for Authorization header

Omit HMAC signature verification on batch message requests.
The OAuth2 authorization server will reject invalid or expired bearer token values before any request reaches NHS Notify.
We only care that a bearer token is present. It is the client's
responsibility to generate a valid token.