Updates for local setup and tidying (#277)
* Update local development docs

To point to the local setup from the main README,
and adding a few helpful notes/corrections.

* add python testing to vscode

this enables the IDE to run tests from UI and
debug breakpoints. I wouldn't normally add to
IDE-specific settings but as they are there
already, and this is a really useful and hopefully
generic setting then it seems ok