Remove CSV file processing function (#276)
* Remove csv file processing function

This function was previously introduced to handle
CSV files for the pilot. There are no plans to use
this in the future and it affects local
development setup, so we are removing it and its
associated tests/fixtures.

* remove send file functions from azurite dependencies

these are no longer used now that we have removed
the function to send csv files.

* Remove azurite setup from e2es

We should no longer need this to run e2es because
the csv file upload function has been removed.

* add a unit test for the migrate function

This has very little value as there is already an
integration test, but it was a learning exercise
for me but does tick a box while I'm looking at
this file anyway.