Require a Consumer when POSTing MessageBatches
This includes renaming the Client model to
Consumer:

I've just dropped the Client table rather than
renaming, as we do not have any remote/production
data in it yet and I'm making a few changes to
columns anyway:

Updating the `key` to be a unique field, as this
will be the human readable string sent by the
service to identify requests.

Removing the `name` and `description` fields until
 we think we might need them.

Making the fk on MessageBatches required, which
means we have to bundle the changes to the
services into this commit.

If a Consumer does not already exist we return a
401, otherwise attach the ID to the MessageBatch.