rename Client to Consumer
This model is going to be used to identify the
senders/origin of message batches - we decided to
make it have a more specific name than Client.

I've just dropped the Client table rather than
renaming, as we do not have any remote/production
data in it yet and I'm making a few changes to
columns anyway:

Updating the `key` to be a unique field, as this
will be the human readable string sent by the
service to identify requests. And removing the
`name` and `description` fields until we think we
might need them.