DTOSS-9044: Require Consumer Key for MessageBatches (#285)
* remove unused imports

* fix import for test_steps

this was throwing error/warning on my IDE

* Add validation helper method for batch POST requests

this will help us test in near future for presence
of the consumer key

* Require a Consumer when POSTing MessageBatches

This includes renaming the Client model to
Consumer:

I've just dropped the Client table rather than
renaming, as we do not have any remote/production
data in it yet and I'm making a few changes to
columns anyway:

Updating the `key` to be a unique field, as this
will be the human readable string sent by the
service to identify requests.

Removing the `name` and `description` fields until
 we think we might need them.

Making the fk on MessageBatches required, which
means we have to bundle the changes to the
services into this commit.

If a Consumer does not already exist we return a
401, otherwise attach the ID to the MessageBatch.

* Update README.md

* constantize the CONSUMER_KEY in headers

this is going to be used a bunch of places so
makes sense to make it a constant