DTOSS-9044: filter Statuses by Consumer (#291)
* Require a `consumer-key` in headers to get statuses

Verify that the headers on GET requests for
statuses includes a consumer key. We will confirm
the Consumer is valid in next commit.

* Require a valid consumer to request statuses

Check the Consumer exists when checking for
Statuses.

* Filter Statuses by Consumer

Here we take the valid Consumer passed in by the
request header and ensure that we only return
statuses for that consumer.

* update README.md

* rename CONSUMER_KEY_NAME to be consistent with other constants