DTOSS-10246: Consisten request headers for Consumer end points (#297)
* add validation for consumer requests

this will be applied to any GET and POST requests
made by a Consumer to e.g. POST Message batches or
GETting Statuses.

* use shared consumer headers for getting statuses

the GET request should now check for authorization
headers as well as the previously checked API key
and Consumer

* remove unused get_statuses_headers

* use standard Consumer headers when POSTing batches

This adds a check for an API Key in the headers,
using the new standard request headers we have
setup for consumers of the API

* remove unused verify_batch_headers