add a flask command to create a new Consumer
this takes the consumer `key` as an argument and
allows us to unit test the behaviour of creating a
new Consumer in the db. This will be used by a
function that can be called on remote environments

can be run locally:

`$ flask --app src/notify/app create-consumer some-key`

I've delegated the actual logic to a separate
method, because we want the option of calling this
in code via a Function. This way we have a command
for ease of use locally, and the ability to create
a Function which will be easier to run with our
current setup.

More info about custom commands
https://flask.palletsprojects.com/en/stable/cli/#custom-commands
https://flask.palletsprojects.com/en/stable/testing/#running-commands-with-the-cli-runner