DTOSS-10207: add a flask command to create a new Consumer (#299)
* Add psycopg2.errors to ignored list for pylint

Pylint was flagging that the UniqueViolation error
did not exist in this module - when I've tested
that the error does exist and it is in the pylint
docs https://www.psycopg.org/docs/errors.html

I tried to add the module to the package
allow-list as suggested here
https://pylint.readthedocs.io/en/latest/user_guide/messages/error/no-member.html
but that did not work, so have gone with suggested
workaround from here
https://github.com/pylint-dev/pylint/issues/2948
as it sounds like it's an issue with psycopg
rather than pylint.

* add a flask command to create a new Consumer

this takes the consumer `key` as an argument and
allows us to unit test the behaviour of creating a
new Consumer in the db. This will be used by a
function that can be called on remote environments

can be run locally:

`$ flask --app src/notify/app create-consumer some-key`

I've delegated the actual logic to a separate
method, because we want the option of calling this
in code via a Function. This way we have a command
for ease of use locally, and the ability to create
a Function which will be easier to run with our
current setup.

More info about custom commands
https://flask.palletsprojects.com/en/stable/cli/#custom-commands
https://flask.palletsprojects.com/en/stable/testing/#running-commands-with-the-cli-runner

* Add Function http request for creating a Consumer

This uses previously created functionality to
create a new Consumer and creates an endpoint for
us to call it with a new `key`. I've tried to
stick as close as possible to conventional REST
behaviour.